remove a circular import
